Correct w definition in density-profile, and cleaning
1. Corrected w as used in density-profile to calculate r(beta) --
I had confused this with omega, the ratio of DE density to matter density.
w is only used in Pavlidou's notes and depends on delta_ta, which she told me
is something like 11 * the matter density, today at least.
    - impliment w(delta_ta), and use this to find the correct analytic density profile.
    - remove Omega_m and Omega_L and use the ones in params. Note this will not be correct
if Omega_m and Omega_L do not take their Lambda-CDM values, as I've needed to use delta_c
and gamma as derived for LCDM in one of Vaso's papers.
    - add comments and clean up axes labels on the plot.

2. In functions, rename mp -> pms and add S(m) function. For now this gives the
power law approximation, in the future I hope to impliment the full definition. Also
divide rho by rho_m today, to give a nice range in the density profile plot.

3. Add sigma_8 (named s_8 here) and comments to parameters.py.

4. Rename mp -> pms in growth-factor.py