Finish double distribution modified
The double distribution program calculates the full double distribution
of number density with respect to mass (smoothing scale) and local
overdensity, as formulated in *Pavlidou and Fields (P&F) 2005.*

The program requires a choice for the **overdensity region parameter beta**,
which I chose to match the results presented in P&F. It then calculates
the double distribution for a set range of masses and overdensities,
again taken to match P&F. This PDF is then calculated in a heatmap with
marginals shown on the axes.

*Note that the plotting routine was mostly
written by Cursor, with small edits from me.*

The double distribution is defined in functions.py. It depends on the
calculation of **a_coll**, the collapse scale factor used to calculate the
universal collapse overdensity delta_c. I chose to calculate a_coll numerically,
by minimising the distance between the integral evaluated at a_coll and the one
evaluated at a_pta, which is known from parameters.

**IMPORTANT:** This code currently only works if the power law approximation
for S(m) is turned off. This is because I don't properly normalise S(m)
in the power law approximation. I have also only tried this
with the concordance set of parameters -- it is possible that trying to use
EdS will break this commit.

Other small changes

functions.py
    - add float requirement for gamma arguments in S, rho and r.
    - remove floor argument in D, use instead the a_i parameter directly.

parameters.py
    - define w (omega), phi and kappa, used to calculate a_coll, for
    each possible cosmology.