Enforce positive constraint in double-distribution
Add option to parameters.py that enforces the joint pdf, calculated
by the dn() function in functions.py, to be positive by zeroing out
values that are negative.

This constraint is enforced directly in dn and so will affect all
calls to this function.

Confirmed that both slices of the pdf and the 2D joint pdf appear
normal with this turned on.