Calculate expectation value of sliced double distribution
Add pdf_expectation to functions, which calculates the expectation
of the double distribution sliced in m. The current implementation
assumes the integration bounds on delta_l are -inft, infty.
As shown, this does not really give the most probable profile, since the
pdf goes negative and pushes the mean further down than it should be.

I also flipped the arguments in func.dn and separated out the calculation
of delta_0_c from dn, as it is also used to find the expectation.