(BUG FIX) Correct PDF norm in calculation of double distribution expectation
(!!) The double distribution as written in previous commits is not properly
normalised.

A normalisation is required to make the double distribution integrate to 1
on the current delta_l range. I implemented the analytic form of this
normalisation in the estimation of the mean from the analytic form (both
derived in Mathematica), and implemented a numerical approximation of the
norm in the NEW sample mean calculation.

I added a sample mean calculation in order to check that the analytic solution
was giving something like the right answer. Both the analytic form of the mean
(blue stars) and the approximation (red dots) are plotted on the resulting slice
plot.