(BUG FIX) Add norm to dn directly and fix sample mean calculation BUG
Normalise the pdf as calculated by dn directly using the analytic form.
Note this will NOT generalise where the floor function is used on the PDF.

Also removed the second loop over mass for calculating the expectations --
this had made the sample expectation incorrect as it only used the PDF[]
list from the final part of the previous loop. It messes up the plot legends,
but surely this is something that can be fixed later.

Also:
- Removed print statements for the sample and analytic norms
- Just food for thought, I don't entirely understand why but it seems the
numerical expectation STILL needs to be normalised on the range, makes me
think I've perhaps done the norm a bit incorrectly here...
- Re-organised plot_dimensions=1 block