Replace sample mean with sample mode
I changed the function calculating the sample statistics on the double
distribution to calculate the mode instead of the mean, as this is
where the most probable profile is really found.

I also changed the sample variance to be centered on the mode, and put both
the mode and variance calculation in the func.pdf_sample_expectation method.

I removed the calculation of the analytic mean, as this is not really the
measure we want.