(IN PROGRESS) Plot PDF w/r/t beta and rho
We would like to plot the PDF with respect to rho (!! normalised by rho_m) and
beta, instead of delta and m, in order to match Fig. 1 of arXiv:2404.18634.
To do this, I've done the following so far in double-distribution.py:

1. Pick out ranges of beta and rho to use, based on Fig. 1 and also that
beta should be >1.3 for sake of the model.
2. Switch spacing so that beta is linearly spaced and rho is log spaced.
3. Adapt the DIM2 block to the new name changes (have not tested this block yet).
4. Adapt the DIM1 block to changes and remove calculation and plot of mode for now.
5. Change the formatting of the DIM1 figure.
6. Comment out the convergence test block for now.

In functions.py, I:

1. Switch the arguments of dn so that beta and rho are input, and m (the virial
mass) is automatically set to M_200 as defined in parameters (see below).
2. Added the calculation of delta_tilde from rho at the beginning, as the
PDF is still written as a function of delta_tilde.
3. Remove the norm for now, this should be re-calculated on paper.
4. Add a part to the norm that converts the PDF into a function of delta (therefore
a function of rho, as rho is just a shift of delta).

Finally, in parameters I replaced the fixed delta_dd with M_200 as used in the
above paper to calculate the most probable profile.