(!BUG!) Completed switch to using rho and beta in double distribution plots
Generated plots of the double dist. w/r/t rho and beta by removing the
norm that translated the derivative from delta_tilde to delta (I guess
this is not needed, still not quite sure why...)

I also reformatted the sample mode calculation function and removed the
normalisation of the stdev calc as this was returning just very large
stdevs.

Note I DO NOT think the normalisation of this PDF is correct in this commit,
the old norm I was using doesn't work anymore (shocker as we changed implicit
variables) and while I tried adding the dS/dm term in... it was just very small.
Plus now m is not the independent variable so hm. Food for thought.