Add possible norm to double distribution and small fixes.
1. I tried adding a norm to the double distribution, according to
the integral of this distribution in Mathematica, but it did not seem
to give the right behavior i.e. I numerically integrated the double distribution
for a max rho of 6, which covers most of the lowest mass double distribution slice,
but this integral was not near 1. So this norm still needs some work.

2. Also changed the plot of the double distribution to have a higher resolution for the
beta and rho ranges, and removed the variance from the double distribution plot as
it seems too large to be realistic.

3. Removed hidden files from the git history, and added a requirements.txt file.