Clean-up and attempt to normalise the double distribution.
1. Added code meant to calculate the appropriate normalisation factor for the double distribution,
according to Mathematica. This did not seem to work, as increasing the rho range to (0, 6) which
gives most of the double distribution at the bottom of the mass range still did not integrate to
something close to 1. I commented the norm out, but would really like to get it working soon.

2. Made delta_c a parameter in parameters.py, as it depends on the cosmology. Fixed references to
delta_c in density-profile.py and double-distribution.py.

3. Removed calculation of the sample variance of the double distribution, as it was much too large, perhaps
due to the bad normalisation.

4. Increased the resolution of the double distribution to 100 steps from 50.

5. Removed unused beta_ta parameter.