Print the accuracy with which we recover the norm
In the diagnostic file, instead of printing the normalisation, which should
always be near 1, I am now printing the accuracy we have on the norm -- that
is, abs(1-norm). This should give one measure of the accuracy at which we can
trust the statistics we derive from the numeric distribution.

Also asked beta to print 4 decimal places, so that the columns format nicely.