Refactor plot_dimension = 1 block
- Loop over the entire range of betas, at each point calculate the PDF, its
sample mode and its analytic prediction for the mode.
- If beta is in the specified range of beta slices, calculate the CDF, normalise
the PDF, print the PDF norm accuracy and plot the PDF slice as well as its mode.

This runs much, much slower as the PDF is now calculated at each point in beta,
should run approximately as quickly as the plot_dimension = 2 block.