(BF) Correct normalisation of the PDF
The 1D PDF was NOT CORRECTLY NORMALISED for previous commits.

I had made exactly the mistake Matt Kenzie warns about in his advanced
stats lectures -- a slice through the PDF IS NOT A PDF and you should
not expect it to be normalised like a PDF.

Essentially, for the 1D plots before I had been norming the PDF
according to the area present just in each slice of beta, but I
should have been using the fully-normalised 2D PDF the entire time.

So, what this commit does:
    1. Move the 2D PDF calculation out of the 2D plot logic block.
    This means every time we run this program we need to calculate the
    full 2D PDF, cause we need it's norm. Sad, it means the program is
    slower. But it is correct!

    2. Correct the norm of the 2D PDF. I had also implimented the 1D
    slice norm here before, very bad. Now, I sum the PDF over the whole
    2D grid and divide it out at the end, before any further plotting
    or calculations are done. I also print this norm precision to terminal,
    it is returning zero now because the norm is very good on the 2D PDF,
    like e-15 ish.

    3. Rename Z -> PDF,

    4. Replace past references to the list PDF with the new PDF 2DArray
    sliced at a particular value of beta,

    5. Move printing mode diffs out of beta_slice block, so this gets
    calc on the whole beta axis. Not sure if I love this yet.

    6. In pdf_sample_expectation, use argmax rather than index(max()) to
    calculate the index at the max, cause now the pdf is an NDArray rather
    than a list.

    7. Move some of the parameters around, cause why not.