Add analytic IQR calculation to double-distribution.py
A. Create function analytic_IQR in double_distribution_functions.py,
which calculates the analytic IQR by finding the root of the CDF at
0.25 and 0.75. Includes print statements used to check the result.
Also moved the PDF sample expectation up as this is used as the initial
guess for the IQR root finder.

- Commented out plot generation in D=1 block of double-distribution.py,
to bug check the IQR calculation.

- minimise rho precision to 0.01, to bug check the IQR calculation quickly.

NOTE this commit is very messy, many things are commented out and print
statements abound. User beware.