Switch back to using the conditional PDF in the dimension=1 block
A. I think the correct slice to calculate the mode and IQR on numerically
may be the conditional PDF, rather than a slice of the normalised 2D PDF.
This is mostly because we can only realistically calculate the conditional
analytic CDF, which is required to get the analytic IQR as derived in
the documentation. Changes on this theme:
- Move 2D calculation / normalisation to the dimension=2D plotting block.
- Calculate the PDF on each beta slice and normalise it there. These two
dramatically reduce comp. time when running the 1D block.
- Plot the conditional PDF, and print only the conditional PDF mode and
stdev from mode to terminal.
- Switch beta_slices from index units to beta units.
- Also comment out printing inside of IQR function, to test the conditional
CDF code.