Normalised double distribution in m/rho domain for dimension=2 block
For the 2D plotting block, switched plot parameters from beta vs rho to
m vs rho, and switched the normalisation to be in the rho/m plane for
each value of the parameter beta.

Now we calculate a 3D grid in beta, rho, m space, but it shouldn't be
too much worse if we keep the resolution in beta fairly small.