Fix bug in mass diff plots and test
A. Fixed the bug in mass diff plots from the renormalisation changes
to the PDF (norming on rho and m plane rather than rho and beta).
This involved just swapping the order in which m and b are given
to the most_probable_rho_transformed function.

B. Testing why the mass difference always goes positive to negative.
Did this by fixing beta = 2 to match the pdf slice plots I had
at the time, and printing the modes I am calculating. This demonstrated
that the most_probable_rho does not actually depend on mass, which
I believe is what leads to the sign flip. But it presents another
mystery because the PDF slice corresponding to this solution does have
a mode which changes with respect to the mass...