Refactor converstion of delta_tilde to/from rho
A. Move calculation of delta_c_0 to functions, along with all of
its helper functions.

B. Add two new functions in functions which convert delta_tilde into
rho and rho into delta_tilde. This was to cut down on duplicated code
in double_distribution_functions.

C. Remove all hard-coded instances of converting delta_tilde and rho
in double_distribution_functions.

D. Move all functions calculating the PDF to the top of
double_distribution_functions, and add more comments detailing the
different sections of this file. Now the layout is:
    1. Functions related to the double distribution ie. the PDF,
    2. Functions related to the CDF of the double distribution, and
    3. Functions related to the IQR, or spread, of this distribution.

E. Uncomment calculation of sample spread from pdf_sample_expectation,
as this will be needed once we move back to looking at the IQR. Also
threw out this value as it is calculated in multi_mass_plots.