Add mass to analytic IQR calculation
Forced the mass to be a provided parameter in the functions that
are used to analytically calculate the IQR. The issue with the IQR
seems to have been that the mass was not set correctly, so the numeric
calculation saw a different slice in mass than the analytic calculation.
This is now fixed, and the IQR estimates match at 1 decimal place
at the smallest mass to all decimal places at the largest mass (where
all of the PDF is seen in the rho range).