Full description of Newton's method module
The motivation for this new module was that the root finders built
into scipy do not support vectorisation, so using those I am forced to
loop over the parameter space of (beta, m) required to replicate Fig. 3
of the mass scaling paper. This is very, very slow, even for a parameter
space of size (25, 3).

So I am designing a very basic module that impliments Newton's method in
a vectorized fashion. The calculation is encapsulated in a class
NewtonsMethod, which has four methods:
1. init, where I read in the section of beta and m parameter space, as
well as the initial guess, gammas and zscore, which are both needed to
calculate the target function.
2. target_fn, where I calculate the difference between the CDF and the
zscores, which is the function I need to minimise,
3. deriv, where I calculate the local derivative of the target function
using a basic two-point downwind template,
4. run, which contains the iteration step.

The converged values are found and stored outside of the iteration,
inside the solutions object. A mask is created to track the indices that
have already been found -- when a new converged value is added to
solutions, the corresponding point in mask is switched from False
to True.

The strategy of the iteration is as follows, once the initial guess (x0)
and the first step (x1) are decided:

1. Compute the second step, which is stored in temp, using Newton's
formula, in a fully-vectorized manner over the parameter space (beta, m).
2. Floor the second step at the min rho, and cieling the second step at
max rho. This prevents oscillations for functions with steep gradients,
as for instance the target function at point (3, 2).
3. Determine the index values which have converged, ie. where the
first digit altered in the second step falls below the tolerance level.
This is done using the argwhere function.
4. For each index in this argwhere array, store the second step in a
separate array which keeps track of the solutions that have converged,
called solutions. Set the mask value at this point in the parameter
space to True.
5. If the entire mask is filled with True, stop the iteration.
6. Loop over the mask, and if the mask value is False, update x0 to x1
and x1 to temp. If not, continue (i.e. do NOT attempt to continue
converging at parameter points that have already converged). Increment
the iterator.
7. Once the iteration step exits, print the solutions object.

This program converges quickly for e.g. (3, 3) sized parameter space.
I am still working to get it to converge for the ideal (25, 3) parameter
space, may need to lower this to (15, 3) as shown in this params.