Cleaning and commenting NewtonsMethod class
Newton's method seemed well-behaved for a parameter space of size
(15, 3), so I am going to run with that for now.

This commit is mostly to add comments, make some of the variable names
more readable, add docstrings and change the final output from the
full array of solutions to the max of the error on the array of roots.

I also added the zscore as a class parameter, and removed the
plotting routine, so if the plot is needed later look for it in the
commit before this.