Re-create Fig. 3 plot with modes and medians
Uncomment the plotting routine in double_distribution, so that the
target figure is generated. Also applied g to NewtonsMethod class, so it
runs with the correct gamma in accordance with the other function in
that loop.

Also fixed a hard coding issue, where in numeric_median_and_IQR I had
hard-coded the shape of the solution array. Made this parameteric
according to the size of beta and gamma axes given in parameters.