(BF) Two bugs fixed and medians vectorised
I finished vectorising the calculation of the medians. During the
process of plotting the medians, I realised that there was a bug, as
the numerical and analytic estimates no longer matched.

This mismatch was due to my bad organisation of the function arguments
for the conditional_CDF method in double_distribution_functions. I had
placed a before gamma, where in most other functions and where I had
called this function in NewtonsMethod.target_fn, I had put gamma
before a. I fixed this by swapping the order of these arguments in
conditional_CDF itself, though future bugs of this form can be fixed
by using the names of a and gamma, instead of their positionality,
as they are defaulted in the function definition.

I also found another bug, where somewhere in NewtonsMethod the input
guess array gets altered, which alters the n_modes array I use as the
initial guess. I fixed this by making the read in of guess within
NewtonsMethod a cbv, not a cbr, with np.copy.

Other stuff:
- Realised it's wasteful to call most of the statistic calculations
within the loop over m, as they automatically run over the entire m
grid now, so moved these outside the m loop. Now the only code inside
the m loop is the plotting and printing routines.
- Added a try/except block to most_probable_rho_transformed to make it
compatible with both scalar and vector input.