Begin porting calculations of double_distribution to separate class
On this branch, I will be re-organising the double_distribution.py file
into two different classes, one for calculations and one for plotting.

This commit starts the double_distribution_calculations file, which
contains the DoubleDistributionCalculations class.

It also ports the calculation of the PDF itself, into the member function
DoubleDistributionCalculations.calc_PDF. Then each place in
double_distribution where this calculation is performed now calls the
class and its member function.

I verified in both the 2D and 1D case that the same plot is produced,
using the member function. This commit only replaces the PDF calculation
performed in the heatmap block and in the slicing_in_rho block, not the
beta scaling block.