Move 2D plot to DoubleDistributionPlots class
Add a new file, double_distribution_plotting, to store
the DoubleDistributionPlots class, which performs all of
the plotting routines.

Moved the 2D plotting routine ONLY to a new member function
of this class, plot_heatmap.

Confirmed this returns the correct plot, when used in place
of the old plotting routine.