Move median and quantile calculation into n_stats function
I decided it would make more sense to calculate all of
the numerical statistics all in one function, rather than
breaking it into mode/variance and quantiles as I had before.

So this version ported the n_quantiles functionality to the
DoubleDistributionCalculations class, but re-programs it
so that the quantiles are found and returned at the end of
n_stats along with the mode and variance.