(!!) Finish moving mode error plotting to DoubleDistribution class
This and the previous commit move the plotting functionality
in the util/multi_mass_plots.py plot into the DoubleDistribution
class, with a new mode chosen by the flag pms.mode_error.

When activated, this routine calculates the difference between the
universal scaling mode with mass dependence (solves a 2nd order
polynomial) and both the correctly transformed analytic mode
and the numerically-calculated mode. These two lines should lie
on top of each other if the analytic transformed mode is being
calculated correctly, with some scatter on the numeric line
due to the binning of the rho axis. This gets better if the rho
resolution is increased, as in this commit.

This plot can be generated for one value of beta, and is generated
for the heuristic beta if a lot of beta values are requested,
or if 4 beta values or less are requested, then this is evaluated
for all values of beta requested.