Add option to relax power law assumption (pla) for S(m)
This commit re-implements the full calculation of S(m)
in terms of the integral of the transfer function, and
adds both:
1. An option to use the full form of S(m) for all plotting
routines in DoubleDistribution, by turning off
pms.power_law_approximation, and
2. An option to compare the sliced PDF in rho between the
PDF which uses the pla and the one that relaxes it.

This second change involved passing the pla flag through
ddc.calc_PDF -> ddfunc.dn -> func.S, and writing a new
option inside of the dd.pdf_slice_in_rho() function.

This commit also vectorises the non-pla option inside of
S(m), as the scipy.integrate.quad function requires
boundaries be scalar-valued only.
The vectorisation was accomplished by
1. Choosing the unique ms passed to the function
(e.g. if ms is a meshgrid in the (beta, rho, m) plane),
2. Creating a dictionary keyed on each unique m, with
values corresponding to the integral evaluated at that m,
and
3. Broadcasting that dictionary into the shape of the
original mesh of ms.
Then the normalisation can be accomplished as normal.

This commit also removes the 4th mode from main.py,
as previous commits have moved this plotting routine
into the DoubleDistribution class.

This commit also changes the logic inside ddp.plot_rho_slice,
so that upon the first call to this function the
defaulted argument primary should be set to True.
This draws the plot in solid lines and saves the mass
slicing color inside of the class. Otherwise, this
argument should not be included, and the secondary plot
is automatically drawn in dashed lines with the matching
mass slicing color palatte.